double > float

+++ 04_simd/13_nbody.cpp
@@ -4,7 +4,7 @@
 
 int main() {
   const int N = 8;
-  double x[N], y[N], m[N], fx[N], fy[N];
+  float x[N], y[N], m[N], fx[N], fy[N];
   for(int i=0; i<N; i++) {
     x[i] = drand48();
     y[i] = drand48();
@@ -14,9 +14,9 @@ int main() {
   for(int i=0; i<N; i++) {
     for(int j=0; j<N; j++) {
       if(i != j) {
-        double rx = x[i] - x[j];
-        double ry = y[i] - y[j];
-        double r = std::sqrt(rx * rx + ry * ry);
+        float rx = x[i] - x[j];
+        float ry = y[i] - y[j];
+        float r = std::sqrt(rx * rx + ry * ry);
         fx[i] -= rx * m[j] / (r * r * r);
         fy[i] -= ry * m[j] / (r * r * r);
       }
